skel/requirements.txt
# Use pip install -r requirements.txt to install
#
# You can install a specific commit from github like this:
# git+git://github.com/scraperwiki/xypath.git@1b387f5

nose
requests==1.2.0
requests-cache==0.4.0
scraperwiki>=0.3.2
python-dateutil>=1.5
